requirements.txt
Django==3.2
djangorestframework==3.12.4
psycopg2-binary
weasyprint
numpy
opencv-python
PyPDF2
PyMuPDF
pdfplumber
python-docx
pandas
openpyxl
openai
gunicorn
whitenoise
python-dotenv==1.0.0
django-storages==1.14.3 
google-cloud-storage
djangorestframework-simplejwt